Refactor: Update import paths in resources and blueprints to reflect new structure
