+++ examples/chalicelib/resources/accounts.py
@@ -4,8 +4,8 @@ from chalice import Response
 
 from agave.core.filters import generic_query
 
+from ...validators import AccountQuery, AccountRequest, AccountUpdateRequest
 from ..models import Account as AccountModel
-from ..validators import AccountQuery, AccountRequest, AccountUpdateRequest
 from .base import app
