--- examples/fastlib/config.py
@@ -1,4 +1,0 @@
-TEST_DEFAULT_USER_ID = 'US123456789'
-TEST_DEFAULT_PLATFORM_ID = 'PT123456'
-TEST_SECOND_USER_ID = 'US987654321'
-TEST_SECOND_PLATFORM_ID = 'PT987654321'
