+++ examples/fastlib/middlewares/authed.py
@@ -4,7 +4,7 @@ from starlette.responses import Response
 from starlette_context import _request_scope_context_storage
 from starlette_context.middleware import ContextMiddleware
 
-from ..config import TEST_DEFAULT_PLATFORM_ID, TEST_DEFAULT_USER_ID
+from ...config import TEST_DEFAULT_PLATFORM_ID, TEST_DEFAULT_USER_ID
 
 
 class AuthedMiddleware(ContextMiddleware):
