+++ examples/fastlib/resources/cards.py
@@ -4,8 +4,8 @@ from fastapi.responses import JSONResponse as Response
 
 from agave.core.filters import generic_query
 
+from ...validators import CardQuery
 from ..models import Card as CardModel
-from ..validators import CardQuery
 from .base import app
