+++ examples/fastlib/resources/files.py
@@ -5,8 +5,8 @@ from fastapi.responses import JSONResponse as Response
 
 from agave.core.filters import generic_query
 
+from ...validators import FileQuery, FileUploadValidator
 from ..models import File as FileModel
-from ..validators import FileQuery, FileUploadValidator
 from .base import app
