+++ examples/fastlib/resources/users.py
@@ -3,8 +3,8 @@ from fastapi.responses import JSONResponse as Response
 
 from agave.core.filters import generic_query
 
+from ...validators import UserQuery, UserUpdateRequest
 from ..models import User as UserModel
-from ..validators import UserQuery, UserUpdateRequest
 from .base import app
