+++ examples/validators.py
@@ -1,3 +1,4 @@
+import datetime as dt
 from typing import Optional
 
 from cuenca_validations.types import QueryParams
@@ -27,6 +28,15 @@ class AccountRequest(BaseModel):
     name: str
 
 
+class AccountResponse(BaseModel):
+    id: str
+    name: str
+    user_id: str
+    platform_id: str
+    created_at: dt.datetime
+    deactivated_at: Optional[dt.datetime] = None
+
+
 class AccountUpdateRequest(BaseModel):
     name: str
 
@@ -37,3 +47,12 @@ class FileQuery(QueryParams):
 
 class CardQuery(QueryParams):
     number: Optional[str] = None
+
+
+class FileUploadValidator(BaseModel):
+    file: bytes
+    file_name: str
+
+
+class UserUpdateRequest(BaseModel):
+    name: str
