+++ tests/core_tests/test_decorators.py
